S3 Buckets, CodeCommit, CodeDeploy, & CodePipeline
The Terraform code is not functional in its current state since I am experiencing an error(s) that appears to be related to the CodeCommit data repository during the Source Stage of the CodePipeline. The status of CodeDeploy and CodePipeline are close to completion, but there is something wrong that is somehow throwing them off. However, the initial S3 Bucket and CodeCommit are working properly when the rest of the code is commented out. The S3 Bucket has to be created in the Ubuntu terminal first, then the repository can be constructed. I have included the associated files such as the dependencies, appspec.yml, and the CodeDeploy installation script for the EC2 instance. I did not include the index.html website file, but it is supposed to be located in the "main" subdirectory within the "Codepipeline" folder.